Working spellbooks
Spellbooks have been tested far enough that I'm ready to merge them back in.  For the moment they generate with multiple spells, but each one can only teach one spell to the player.

I've also added a victory condition (getting deep enough), enlarged the map size, and turned off the debug vision, so it is in 'play it like it's a real game' mode.